+++ src/modules/treasury/services/transaction.service.ts
@@ -61,12 +61,15 @@ export class TransactionService {
     id: string,
     transaction: Partial<Transaction>,
     userId: string,
-  ): Promise<Transaction> {
+  ): Promise<Transaction | null> {
     const existingTransaction = await this.transactionRepository.findById(id);
     const updatedTransaction = await this.transactionRepository.update(
       id,
       transaction,
     );
+
+    if (!existingTransaction || !updatedTransaction)
+      throw new Error('existing/updated transaction not found');
 
     // If the status changed to completed, update the asset
     if (
@@ -96,6 +99,7 @@ export class TransactionService {
 
   async delete(id: string, userId: string): Promise<void> {
     const existingTransaction = await this.transactionRepository.findById(id);
+    if (!existingTransaction) throw new Error('existing transaction not found');
     await this.transactionRepository.delete(id);
 
     // Log the delete action
@@ -116,6 +120,7 @@ export class TransactionService {
     if (!transaction.assetId) return;
 
     const asset = await this.assetRepository.findById(transaction.assetId);
+    if (!asset) throw new Error('asset not found');
 
     let newAmount = Number(asset.amount);
